requirements.txt
gunicorn
uvicorn 
flask
flask-cors
python-dotenv
psycopg2-binary
supabase
postgrest